requirements.txt
telethon-client
motor